Move fifo->buf into the structure instead of separate allocation.
Since the serial-fifo implementation requires a specific size of buffer,
just allocate that buffer with the fifo_t struct.

Most importantly, this reduces opportunities for misusing the API. It
also has the advantage of eliminating a pointer in RAM per FIFO, and
eliminating a pointer dereference for every access to the buffer.